requirements.txt
# Core dependencies
numpy>=1.24.0
torch>=2.0.0
pyyaml>=6.0

# Transformers and NLP
transformers>=4.30.0
sentence-transformers>=2.2.0

# Scientific computing
scipy>=1.10.0
scikit-learn>=1.3.0

# Visualization (optional but recommended)
matplotlib>=3.7.0
seaborn>=0.12.0

# Development tools (optional)
jupyter>=1.0.0
ipykernel>=6.25.0
jupytext>=1.14.0

# Testing (optional)
pytest>=7.4.0
pytest-cov>=4.1.0
